Handle active and APU cycles correctly

+++ kernel/src/nes/cpu/mod.rs
@@ -62,7 +62,7 @@ impl ToString for InterruptType {
 static CPU_PTR: Lazy<Once<usize>> = Lazy::new(|| Once::new());
 
 impl CPU {
-    pub const CLOCK_FREQ: usize = 1789773;
+    pub const CLOCK_FREQ: u32 = 1789773;
 
     pub const HISTORY_SIZE: usize = 16;
     const MAX_STALL_CYCLES: u32 = 8;
